Move C++ tests
