+++ source/neuropod/backends/python_bridge/test/gpu_test_python_bridge.cc
@@ -13,16 +13,28 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "test_utils.hh"
+#include "neuropod/tests/test_utils.hh"
+
+#include <thread>
+
+#include <patchlevel.h>
 
 TEST(test_models, test_pytorch_addition_model)
 {
+#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8
     // Test the PyTorch addition model using the python bridge
-    test_addition_model("neuropod/tests/test_data/pytorch_addition_model/");
+    test_addition_model("neuropod/tests/test_data/pytorch_addition_model_gpu/");
+#endif
 }
 
-TEST(test_models, test_pytorch_strings_model)
+TEST(test_models, test_pytorch_addition_model_threaded)
 {
-    // Test the PyTorch strings model using the python bridge
-    test_strings_model("neuropod/tests/test_data/pytorch_strings_model/");
+#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8
+    std::thread t([]() {
+        // Test the PyTorch addition model using the python bridge
+        test_addition_model("neuropod/tests/test_data/pytorch_addition_model_gpu/");
+    });
+
+    t.join();
+#endif
 }
