+++ source/neuropod/backends/python_bridge/test/test_python_bridge.cc
@@ -13,28 +13,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "test_utils.hh"
-
-#include <thread>
-
-#include <patchlevel.h>
+#include "neuropod/tests/test_utils.hh"
 
 TEST(test_models, test_pytorch_addition_model)
 {
-#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8
     // Test the PyTorch addition model using the python bridge
-    test_addition_model("neuropod/tests/test_data/pytorch_addition_model_gpu/");
-#endif
+    test_addition_model("neuropod/tests/test_data/pytorch_addition_model/");
 }
 
-TEST(test_models, test_pytorch_addition_model_threaded)
+TEST(test_models, test_pytorch_strings_model)
 {
-#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8
-    std::thread t([]() {
-        // Test the PyTorch addition model using the python bridge
-        test_addition_model("neuropod/tests/test_data/pytorch_addition_model_gpu/");
-    });
-
-    t.join();
-#endif
+    // Test the PyTorch strings model using the python bridge
+    test_strings_model("neuropod/tests/test_data/pytorch_strings_model/");
 }
