+++ source/neuropod/backends/tensorflow/test/BUILD
@@ -1,4 +1,4 @@
-# Copyright (c) 2020 UATC, LLC
+# Copyright (c) 2021 UATC, LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,18 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-cc_library(
-    name = "neuropod_test_utils",
-    testonly = True,
-    hdrs = ["test_utils.hh"],
-    data = [
-        "//neuropod/tests/test_data",
-    ],
-    visibility = [
-        "//visibility:public",
+cc_test(
+    name = "test_tensorflow_backend",
+    srcs = [
+        "test_tensorflow_backend.cc",
     ],
     deps = [
         "//neuropod:neuropod_impl",
-        "@gtest//:main",
+        "//neuropod/backends/tensorflow:tensorflow_backend",
+        "//neuropod/tests:neuropod_test_utils",
     ],
 )
