+++ source/neuropod/bindings/c/test/BUILD
@@ -1,4 +1,4 @@
-# Copyright (c) 2020 UATC, LLC
+# Copyright (c) 2021 UATC, LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,18 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-cc_library(
-    name = "neuropod_test_utils",
-    testonly = True,
-    hdrs = ["test_utils.hh"],
+cc_test(
+    name = "test_c_api",
+    srcs = [
+        "test_c_api.c",
+    ],
+    copts = ["-std=c11"],
     data = [
         "//neuropod/tests/test_data",
     ],
-    visibility = [
-        "//visibility:public",
-    ],
     deps = [
         "//neuropod:neuropod_impl",
+        "//neuropod/backends/tensorflow:tensorflow_backend",
+        "//neuropod/bindings/c:c_api",
         "@gtest//:main",
     ],
 )
